part-3: sync with Nabin's

+++ FitFromHist/makeAllDCs.py
@@ -3,9 +3,10 @@ import itertools
 import json
 from FitInputs import *
 
-with open("DataCards.json", "w") as f:
-    data = {}
-    json.dump(data, f)
+if not os.path.exists("./DataCards.json"):
+    with open("DataCards.json", "w") as f:
+        data = {}
+        json.dump(data, f)
 
 def runCmd(cmd):
     print "\n\033[01;32m Excecuting: %s \033[00m"%cmd
@@ -16,14 +17,14 @@ for year, decay, channel in itertools.product(Year, Decay, Channel):
     #runCmd("python makeDCInc.py %s"%args)
     runCmd("python makeDCCat.py --isChIsoM3      --hist phosel_noCut_ChIso %s"%args)
     runCmd("python makeDCCat.py --isChIsoM3      --hist phosel_M3 %s"%args)
-    runCmd("python makeDCCat.py --is0PhoM3       --hist presel_M3 %s"%args)
+    runCmd("python makeDCCat.py --is0PhoM3       --hist presel_M3_0Pho %s"%args)
+    runCmd("python makeDCCat.py --isMassLepGamma --hist phosel_MassLepGamma %s"%args)
     runCmd("python makeDCCat.py --isMassDilep    --hist presel_MassDilep %s"%args)
-    runCmd("python makeDCCat.py --isMassLepGamma --hist phosel_MassLepGammma %s"%args)
     for CR in ControlRegion:
         args = "-y %s -d %s -c %s --cr %s"%(year, decay, channel, CR)
         #runCmd("python makeDCInc.py %s"%args)
         runCmd("python makeDCCat.py --isChIsoM3      --hist phosel_noCut_ChIso %s"%args)
         runCmd("python makeDCCat.py --isChIsoM3      --hist phosel_M3 %s"%args)
-        runCmd("python makeDCCat.py --is0PhoM3       --hist presel_M3 %s"%args)
+        runCmd("python makeDCCat.py --is0PhoM3       --hist presel_M3_0Pho %s"%args)
+        runCmd("python makeDCCat.py --isMassLepGamma --hist phosel_MassLepGamma %s"%args)
         runCmd("python makeDCCat.py --isMassDilep    --hist presel_MassDilep %s"%args)
-        runCmd("python makeDCCat.py --isMassLepGamma --hist phosel_MassLepGammma %s"%args)
